requirements.txt
discord.py[voice]>=2.3.2
yt-dlp>=2023.11.16
PyNaCl>=1.5.0
python-dotenv>=1.0.0 